requirements.txt
Flask
gunicorn
requests
pandas
pandas-ta
yfinance
mplfinance
Pillow